// WebApp.Client/src/generated/dtos.ts
export class ProblemDetails implements IProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;

    constructor(data?: IProblemDetails) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.type = _data["type"];
            this.title = _data["title"];
            this.status = _data["status"];
            this.detail = _data["detail"];
            this.instance = _data["instance"];
        }
    }

    static fromJS(data: any): ProblemDetails {
        data = typeof data === 'object' ? data : {};
        let result = new ProblemDetails();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["type"] = this.type;
        data["title"] = this.title;
        data["status"] = this.status;
        data["detail"] = this.detail;
        data["instance"] = this.instance;
        return data;
    }
}

export interface IProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;
}
// ...
export class UserMinimalDto implements IUserMinimalDto {
    id?: number;
    userName?: string;

    constructor(data?: IUserMinimalDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.userName = _data["userName"];
        }
    }

    static fromJS(data: any): UserMinimalDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserMinimalDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["userName"] = this.userName;
        return data;
    }
}

export interface IUserMinimalDto {
    id?: number;
    userName?: string;
}
// ...
export class UserHeaderDto extends UserMinimalDto implements IUserHeaderDto {
    email?: string;
    dateOfBirth?: Date;
    totalBalance?: number;

    constructor(data?: IUserHeaderDto) {
        super(data);
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.email = _data["email"];
            this.dateOfBirth = _data["dateOfBirth"] ? new Date(_data["dateOfBirth"].toString()) : undefined as any;
            this.totalBalance = _data["totalBalance"];
        }
    }

    static override fromJS(data: any): UserHeaderDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserHeaderDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["dateOfBirth"] = this.dateOfBirth ? this.dateOfBirth.toISOString() : undefined as any;
        data["totalBalance"] = this.totalBalance;
        super.toJSON(data);
        return data;
    }
}

export interface IUserHeaderDto extends IUserMinimalDto {
    email?: string;
    dateOfBirth?: Date;
    totalBalance?: number;
}
// ...
export class UpdateUserEmailDto implements IUpdateUserEmailDto {
    email?: string;

    constructor(data?: IUpdateUserEmailDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.email = _data["email"];
        }
    }

    static fromJS(data: any): UpdateUserEmailDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateUserEmailDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        return data;
    }
}

export interface IUpdateUserEmailDto {
    email?: string;
}

export class UpdateUserNameDto implements IUpdateUserNameDto {
    userName?: string;

    constructor(data?: IUpdateUserNameDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.userName = _data["userName"];
        }
    }

    static fromJS(data: any): UpdateUserNameDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateUserNameDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userName"] = this.userName;
        return data;
    }
}

export interface IUpdateUserNameDto {
    userName?: string;
}

export class UpdateUserDateOfBirthDto implements IUpdateUserDateOfBirthDto {
    dateOfBirth?: Date;

    constructor(data?: IUpdateUserDateOfBirthDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.dateOfBirth = _data["dateOfBirth"] ? new Date(_data["dateOfBirth"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): UpdateUserDateOfBirthDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateUserDateOfBirthDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["dateOfBirth"] = this.dateOfBirth ? this.dateOfBirth.toISOString() : undefined as any;
        return data;
    }
}

export interface IUpdateUserDateOfBirthDto {
    dateOfBirth?: Date;
}

export class UpdateUserPasswordDto implements IUpdateUserPasswordDto {
    oldPassword?: string;
    newPassword?: string;

    constructor(data?: IUpdateUserPasswordDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.oldPassword = _data["oldPassword"];
            this.newPassword = _data["newPassword"];
        }
    }

    static fromJS(data: any): UpdateUserPasswordDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateUserPasswordDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["oldPassword"] = this.oldPassword;
        data["newPassword"] = this.newPassword;
        return data;
    }
}

export interface IUpdateUserPasswordDto {
    oldPassword?: string;
    newPassword?: string;
}